+++ pennyway-client-iOS/pennyway-client-iOS/Api/Domain/UserDomain/Alamofire/UserAuthAlamofire.swift
@@ -31,10 +31,4 @@ class UserAuthAlamofire: TokenHandler {
         
         ApiRequstHandler.shared.requestWithErrorHandling(session: session, router: UserAuthRouter.checkLoginState, completion: completion)
     }
-    
-    func refresh(completion: @escaping (Result<Data?, Error>) -> Void) {
-        Log.info("UserAuthAlamofire - refresh() called")
-        
-        ApiRequstHandler.shared.requestWithErrorHandling(session: session, router: UserAuthRouter.refresh, completion: completion)
-    }
 }
